Add example code + various fixes.

+++ include/evdevw/ReadFlag.hpp
@@ -1,6 +1,8 @@
 #ifndef EVDEVW_READFLAG_HPP
 #define EVDEVW_READFLAG_HPP
 
+#include <cstdint>
+
 #include "Enum.hpp"
 
 namespace evdevw {
@@ -13,13 +15,13 @@ namespace evdevw {
   };
 
   template <>
-  int enum_to_raw<int, ReadFlag>(ReadFlag code) {
+  uint16_t enum_to_raw<uint16_t, ReadFlag>(ReadFlag code) {
     using UT = std::underlying_type_t<ReadFlag>;
     return static_cast<UT>(code);
   }
 
   template <>
-  ReadFlag raw_to_enum<ReadFlag, int>(int code) {
+  ReadFlag raw_to_enum<ReadFlag, uint16_t>(uint16_t code) {
     return static_cast<ReadFlag>(code);
   }
